Completely reworked parsing loop to be concurrent.
Each file is now parsed concurrently in a separate go routine
in a line by line fashion. This approach consumes significantly
less memory compared to the previous in memory approach since only
a single line of each file is read at a time. Thus, very large files
can now be processed. Please note that pst will open all file provided
on the command line simultaneously and may thus fail if called for large
numbers of files depending on the OS and the limit on max number of
simultaneously open files.