Fixed bug in assembling output columns.
The previous code would combine the parsed columns for each
input file into a single string before passing it down the
data channel. This of course doesn't work together
with the flexible output column specification since it does
not allow for re-ordering of columns within and in particular
between files.